changed the iterator count variable to use a smaller integer.  Giving more arguments than that integer can hold results in infinite loops.

+++ ex10/ex10.c
@@ -1,8 +1,9 @@
 #include <stdio.h>
+#include <stdint.h>
 
 int main(int argc, char *argv[])
 {
-  int i = 0;
+  uint16_t i = 0;
 
   /* go through each string in argv.  Skipping argv[0] because it
    contains the program's name. */
